backend/requirements.txt
flask
flask-cors
flask-restful
flask-sqlalchemy
pandas
numpy
requests
gunicorn
